Crate burn_common

Burn Common Library

This library contains common types used by other Burn crates that must be shared.

Re-exports

pub use rayon;

Modules

benchmark

Module for benchmarking any executable part

future

Future utils with a compatible API for native, non-std and wasm environments.

id

Id module contains types for unique identifiers.

parallel

Parallel utilities.

rand

Rand module contains types for random number generation for non-std environments and for std environments.

reader

Useful when you need to read async data without having to decorate each function with async notation.

stub

Stub module contains types for stubs for non-std environments and for std environments.

Macros

iter_par

Macro for iterating in parallel.

iter_range_par

Macro for iterating over a range in parallel.

iter_slice_par

Macro for iterating in parallel.

run_par

Macro for running a function in parallel.