burn_common

Module rand

Rand module contains types for random number generation for non-std environments and for std environments.

Structs

StdRng

The standard RNG. The PRNG algorithm in StdRng is chosen to be efficient on the current platform, to be statistically strong and unpredictable (meaning a cryptographically secure PRNG).

Traits

Rng

An automatically-implemented extension trait on RngCore providing high-level generic methods for sampling values and other convenience methods.

SeedableRng

A random number generator that can be explicitly seeded.

Functions

gen_random

Generates random data from a thread-local RNG.

get_seeded_rng

Returns a seeded random number generator using entropy.